V2 blueprint, static done
