+++ admin_test.go
@@ -1,24 +1,30 @@
-package main
+package gadmin
 
 import (
 	"database/sql"
-	"gadmin/api"
+	"testing"
 	"time"
 
 	"github.com/glebarez/sqlite"
+	"github.com/stretchr/testify/assert"
 	"gorm.io/gorm"
+	"gorm.io/gorm/clause"
 	"gorm.io/gorm/logger"
 	"gorm.io/gorm/schema"
 )
 
-func main() {
-	db, _ := gorm.Open(sqlite.Open("../db.sqlite"),
+func TestApi(t *testing.T) {
+	is := assert.New(t)
+
+	db := must[*gorm.DB](gorm.Open(sqlite.Open("db.sqlite"),
 		&gorm.Config{
 			NamingStrategy: schema.NamingStrategy{SingularTable: true},
 			Logger:         logger.Default.LogMode(logger.Info),
-		})
+		}))
 
-	A := api.NewAdmin("Test Site", db)
+	_ = clause.Associations
+
+	A := NewAdmin("Test Site", db)
 
 	type Foo struct {
 		ID           uint `gorm:"primaryKey"`
@@ -33,8 +39,14 @@ func main() {
 		CreatedAt    time.Time `gorm:"autoCreateTime"`
 		UpdatedAt    time.Time `gorm:"autoUpdateTime:nano"`
 	}
+	fv := NewModelView(Foo{})
+	is.Len(fv.GetBlueprint().Children, 9)
 
-	A.AddView(api.NewModelView(Foo{}))
+	is.Equal("/foo/", fv.GetUrl(".index_view"))
+	is.Equal("/foo/action", fv.GetUrl(".action_view"))
+	is.Equal("/foo/action?a=b", fv.GetUrl(".action_view", "a", "b"))
+
+	A.AddView(fv)
 
 	// belongs to https://gorm.io/docs/belongs_to.html
 	type Company struct {
@@ -47,8 +59,8 @@ func main() {
 		CompanyId int
 		Company   *Company
 	}
-	A.AddView(api.NewModelView(Company{}, "Association"))
-	A.AddView(api.NewModelView(Employee{}, "Association"))
+	A.AddView(NewModelView(Company{}, "Association"))
+	A.AddView(NewModelView(Employee{}, "Association"))
 
 	// has one https://gorm.io/docs/has_one.html
 	type CreditCard struct {
@@ -60,8 +72,8 @@ func main() {
 		gorm.Model
 		CreditCard CreditCard
 	}
-	A.AddView(api.NewModelView(CreditCard{}, "Association"))
-	A.AddView(api.NewModelView(User{}, "Association"))
+	A.AddView(NewModelView(CreditCard{}, "Association"))
+	A.AddView(NewModelView(User{}, "Association"))
 
 	// has many https://gorm.io/docs/has_many.html
 	type Address struct {
@@ -73,8 +85,8 @@ func main() {
 		gorm.Model
 		Addresses []Address
 	}
-	A.AddView(api.NewModelView(Address{}, "Association"))
-	A.AddView(api.NewModelView(Account{}, "Association"))
+	A.AddView(NewModelView(Address{}, "Association"))
+	A.AddView(NewModelView(Account{}, "Association"))
 
 	// many to many https://gorm.io/docs/many_to_many.html
 	type Language struct {
@@ -85,8 +97,8 @@ func main() {
 		gorm.Model
 		Languages []Language `gorm:"many2many:student_language"`
 	}
-	A.AddView(api.NewModelView(Language{}, "Association"))
-	A.AddView(api.NewModelView(Student{}, "Association"))
+	A.AddView(NewModelView(Language{}, "Association"))
+	A.AddView(NewModelView(Student{}, "Association"))
 
 	// polymorphic https://gorm.io/docs/polymorphism.html
 	type Toy struct {
@@ -100,8 +112,8 @@ func main() {
 		Name string
 		Toys []Toy `gorm:"polymorphic:Owner"`
 	}
-	A.AddView(api.NewModelView(Toy{}, "Association"))
-	A.AddView(api.NewModelView(Dog{}, "Association"))
+	A.AddView(NewModelView(Toy{}, "Association"))
+	A.AddView(NewModelView(Dog{}, "Association"))
 
-	A.Run()
+	is.Equal("/admin/foo/", A.UrlFor("foo.index"))
 }
