+++ examples/simple/main.go
@@ -1,8 +1,13 @@
 package main
 
 import (
-	"gadmin"
 	"time"
+
+	"github.com/glebarez/sqlite"
+	"github.com/pedia/gadmin"
+	"gorm.io/gorm"
+	"gorm.io/gorm/logger"
+	"gorm.io/gorm/schema"
 
 	"github.com/google/uuid"
 )
@@ -42,11 +47,17 @@ type Tag struct {
 }
 
 func main() {
-	admin := gadmin.NewAdmin("Example: Simple Views")
-	admin.AddView(gadmin.NewView("Test", "view1"))
-	admin.AddView(gadmin.NewView("Test", "view2"))
+	db, _ := gorm.Open(sqlite.Open("db.sqlite"),
+		&gorm.Config{
+			NamingStrategy: schema.NamingStrategy{SingularTable: true},
+			Logger:         logger.Default.LogMode(logger.Info),
+		})
 
-	mv := gadmin.NewModalView(User{}).
+	admin := gadmin.NewAdmin("Example: Simple Views", db)
+	// admin.AddView(gadmin.NewView(gadmin.MenuItem{Category: "Test", Name: "View1"}))
+	// admin.AddView(gadmin.NewView(gadmin.MenuItem{Category: "Test", Name: "View2"}))
+
+	mv := gadmin.NewModelView(User{}).
 		SetColumnList("type", "first_name", "last_name", "email", "ip_address", "currency", "timezone", "phone_number").
 		SetColumnEditableList("first_name", "type", "currency", "timezone").
 		SetColumnDescriptions(map[string]string{"first_name": "名"}).
@@ -54,7 +65,7 @@ func main() {
 		SetPageSize(5).
 		SetTablePrefixHtml(`<h4>hello</h4>`)
 	admin.AddView(mv)
-	admin.AddView(gadmin.NewModalView(Post{}))
-	admin.AddView(gadmin.NewModalView(Tag{}))
+	admin.AddView(gadmin.NewModelView(Post{}))
+	admin.AddView(gadmin.NewModelView(Tag{}))
 	admin.Run()
 }
